Add `CUDA_HOME` to the `PATH`
To ensure that the CUDA binaries are on the `PATH`, add them to the
`PATH`. As we overwrite the `PATH` before this point, we cannot add the
binaries to the `PATH` earlier (like in the Docker image) as they will
be overwritten. So add the CUDA binaries in the `entrypoint_source`
script.