Drop `cudnn` from the container's cached packages
As we do not have CUDA 10.2 package of `cudnn`, drop this from the
cached packages in the container so that we are able to build this
image. We may decide to add this back later due to the size of the
package and how long it takes to download. However this will at least
help us make forward progress in the near term and we can always
reevaluate this decision in the future.